pyproject.toml
[project]
name = "actionable-finance"
version = "0.1.0"
description = "Telegram bot with webhook support"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "telethon>=1.37.0",
    "pyupbit>=0.2.33",
    "python-binance>=1.0.19",
    "mojito2>=0.1.0",
    "apscheduler>=3.10.0",
    "mplfinance>=0.12.9",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "pykrx>=1.0.51",
    "setuptools>=80.10.2",
    "finance-datareader>=0.9.50",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]
